generation_testing/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
huggingface_hub>=0.20.3
python-multipart>=0.0.6
numpy>=1.24.0
torch>=2.0.0
pydantic>=2.5.0
starlette>=0.36.0
pillow>=10.0.0
requests>=2.31.0 